feat : 데이터 패치 컴포넌트 수정
